requirements.txt
Flask==0.12.2
Flask-JWT-Extended==3.7.2
Flask-RESTful==0.3.6
Flask-Cors==3.0.3
json-schema-generator==0.3
jsonschema==2.6.0
pymongo==3.6.0
python-dotenv==0.7.1
python-json-logger==0.1.8
schema==0.6.7
psycopg2-binary